move structs from decoding.h to appropriate headers, remove some unimplemented/unused functions
