+++ nvse/nvse/GameTypes.h
@@ -717,3 +717,116 @@ struct CoordXY {
 		return *this;
 	}
 };
+
+#define DECL_FLOAT_OP(op) \
+	NiPoint3 operator op(const float n) const \
+	{ \
+		return NiPoint3(x op n, y op n, z op n); \
+	} \
+	NiPoint3 operator op##=(const float n) \
+	{ \
+		return *this = NiPoint3(x op n, y op n, z op n); \
+	} \
+
+#define DECL_VEC_OP(op) \
+	NiPoint3 operator op(const NiPoint3 v) const \
+	{ \
+		return NiPoint3(x op v.x, y op v.y, z op v.z); \
+	} \
+	NiPoint3 operator op##=(const NiPoint3 v) \
+	{ \
+		return *this = NiPoint3(x op v.x, y op v.y, z op v.z); \
+	}
+
+struct NiPoint3 {
+	float x, y, z;
+
+	void Scale(float scale) {
+		x *= scale;
+		y *= scale;
+		z *= scale;
+	};
+
+	void Init(NiPoint3* point) {
+		x = point->x;
+		y = point->y;
+		z = point->z;
+	};
+
+	NiPoint3() : x(0.f), y(0.f), z(0.f) {};
+
+	NiPoint3(const float x, const float y, const float z) : x(x), y(y), z(z) {};
+
+	DECL_FLOAT_OP(*);
+	DECL_FLOAT_OP(/ );
+
+	DECL_VEC_OP(+);
+	DECL_VEC_OP(-);
+	DECL_VEC_OP(*);
+	DECL_VEC_OP(/ );
+
+	float length() const {
+		return sqrtf(x * x + y * y + z * z);
+	}
+
+	float length_sqr() const {
+		return x * x + y * y + z * z;
+	}
+
+	NiPoint3 normal() const {
+		const auto len = length();
+		return len == 0.F ? NiPoint3() : NiPoint3(x / len, y / len, z / len);
+	}
+
+	static float dot(const NiPoint3& v1, const NiPoint3& v2) {
+		return v1.x * v2.x + v1.y * v2.y + v1.z * v2.z;
+	}
+	inline float Distance(const NiPoint3& pt) const {
+		return (*this - pt).length();
+	}
+
+	static NiPoint3 cross(const NiPoint3& v1, const NiPoint3& v2) {
+		return NiPoint3(
+			v1.y * v2.z - v1.z * v2.y,
+			v1.z * v2.x - v1.x * v2.z,
+			v1.x * v2.y - v1.y * v2.x);
+	}
+
+	NiPoint3* Add(NiPoint3* toAdd) {
+		this->x += toAdd->x;
+		this->y += toAdd->y;
+		this->z += toAdd->z;
+		return this;
+	}
+
+	NiPoint3* Subtract(NiPoint3* point) {
+		this->x -= point->x;
+		this->y -= point->y;
+		this->z -= point->z;
+		return this;
+	}
+
+	float CalculateDistSquared(NiPoint3* to) {
+		float deltaX = (x - to->x);
+		float deltaY = (y - to->y);
+		float deltaZ = (z - to->z);
+
+		return deltaX * deltaX + deltaY * deltaY + deltaZ * deltaZ;
+	}
+	static float Sign(NiPoint3 p1, NiPoint3 p2, NiPoint3 p3) {
+		return (p1.x - p3.x) * (p2.y - p3.y) - (p2.x - p3.x) * (p1.y - p3.y);
+	}
+
+	static bool PointInTriangle(NiPoint3 pt, NiPoint3 v1, NiPoint3 v2, NiPoint3 v3) {
+		bool b1 = Sign(pt, v1, v2) < 0.0;
+		bool b2 = Sign(pt, v2, v3) < 0.0;
+		bool b3 = Sign(pt, v3, v1) < 0.0;
+
+		return (b1 == b2) && (b2 == b3);
+	}
+
+	static NiPoint3 GetTriangleCenter(NiPoint3 v1, NiPoint3 v2, NiPoint3 v3) {
+		return NiPoint3((v1.x + v2.x + v3.x) / 3.0f, (v1.y + v2.y + v3.y) / 3.0f, (v1.z + v2.z + v3.z) / 3.0f);
+	}
+
+};
